requirements.txt
rpa~=1.50.0
tagui~=1.50.0
python-dotenv~=1.0.1